Run from mounted volume
There are errors in trusted advisor around the root file system being
writeable.
I've made two new volumes, one for /tests/boot and one for /temp.
When the image runs, it copies the files from the existing /test
directory and runs them in there. I've overidden user.home to get sbt to
write it's cache files there instead of in /root/.sbt
For some reason, it insists on using the /tmp directory so I've mounted
a volume for that.

The PR for the volumes is https://github.com/nationalarchives/dr2-terraform-environments/pull/56